Use longer password for key generation
Use longer password, "dummy", instead of "a", for key generation with
`make keys`.

When running `make keys`, I see error messages of the form

  routines:UI_set_result_ex:result too small:crypto/ui/ui_lib.c:905:You
  must type in 4 to 1023 characters

and `private-key.pem` is generated with a length of 0 bytes.

According to https://bugzilla.redhat.com/show_bug.cgi?id=1467669 ,
this is because the used passwords are too short.