Use larger key size in Makefile
With the default key size of 1024, when I start the server with
`make compile run`, I see several stacktraces with the message

  #<thread:'package-change-handler>
  *** DAEMON CRASHED: main-web-server-thread ***
  ssl-load-certificate-chain!: load failed from: #<path:/home/schwa/sd/racket/racket-pkg-website/src/../server-cert.pem> (error:140AB18F:SSL routines:SSL_CTX_use_certificate:ee key too small)
    context...:

https://github.com/jfromaniello/selfsigned/issues/33 suggests increasing
the key size to at least 2048 bits. Increasing the key size in the
Makefile makes the above stacktraces go away.

I set the key size in the Makefile to 4096 to make it more
"future-proof", but feel free to use a key size of 2048, which works for
me as well.