merge `pkg-index` back end, unify and streamline configuration
Running the server through `run` (or `make run` or `racket src/main.rkt`)
runs both the website front-end server and the pkg-index back end server.

Change the configration table to derive website and pkg-index
configuration from shared keys, so they are in sync by default, and
change all defaults to make sense for testing mode. Also, change
defaults to be relative to the directory where the server is started,
which doesn't have to be the source directory; make the default
configuration put generated files in "compiled", so they're clearly
and reliably distinguished from non-generated files.