adjust back end to force on periodic scanning
The back end schedules a checksum scan every hour, but also skips a
checksum scan for a package if it was checked within the last hour.
The consequence of that combination is that the detection of new
checksums can be delayed up to two hours, instead of the intended 1
hour, and the scan may report a heartbeat even when it hasn't looked
at some packages. To avoid that confusing behavior, c hange the
periodic scan to ignore the last-scanned time.